moved functions to new models
